+++ extension/esbuild.js
@@ -14,7 +14,7 @@ async function main() {
     platform: 'node',
     outfile: 'dist/extension.js',
     external: ['vscode'],
-    logLevel: 'silent',
+    logLevel: 'info',
     plugins: [
       /* add to the end of plugins array */
       esbuildProblemMatcherPlugin,